feat: separate MCC-exclusive endpoints to different package
